+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This is a bug fix release for various issues discovered after we released v0.2.8
 * FEATURE: sections can be grouped as a "view" with the markup `\view{project: id1,id2,idN}`. The grouped sections can be viewed on their own drawer.
 * FEATURE: sections can be pinned for fast access. Pinned sections can be viewed on their own drawer.
 * FEATURE: search results are rendered in their own drawer and search queries can also be performed across all projects.
+* FEATURE: sections can be exported to PDF with `typst`.
 * OPTIMIZED: changed color of links and inline code text.
 * OPTIMIZED: view markup can now accept either chapter name or section IDs.
 * FIXED: fixed bugs associated with rendering lists.